+++ wrapper-ripple/src/main/java/com/hatiko/ripple/telegram/bot/database/rest/controller/message_id/GetMessageIdController.java
@@ -21,7 +21,7 @@ public class GetMessageIdController {
 	private final XrpDatabaseOperator databaseOperator;
 	
 	@GetMapping
-	public ResponseEntity<MessageIdDTO> getMessageId(@RequestParam("chat_id") Integer chatId){
+	public ResponseEntity<MessageIdDTO> getMessageId(@RequestParam("chat_id") Long chatId){
 
 		MessageIdDTO response = databaseOperator.getMessageId(chatId);
 		return ResponseEntity.ok(response);
